The salt logging mix-ins now have it's own module.
